pyproject.toml
[project]
name = "image-token"
version = "0.1.4"
description = "A package to calculate the number of tokens in an image or folder"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.9"
dependencies = [
    "pytest>=8.3.5",
    "pillow (>=11.2.1,<12.0.0)",
    "langchain-openai (>=0.3.18,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "ruff (>=0.11.11,<0.12.0)",
    "requests (>=2.32.4,<3.0.0)",
    "google-genai (>=1.0.0,<2.0.0)"
]

[project.urls]
Homepage = "https://github.com/srinathmkce/imagetoken"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["image_token"]